+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * `[progress]` Code refactoring
 * `[req]` Code refactoring
 * `[spellcheck]` Code refactoring
+* `[tmp]` Code refactoring
 * `[usage]` Code refactoring
 * `[cache]` Better tests for panics
 * `[cron]` Better tests for panics
@@ -23,6 +24,7 @@
 * `[progress]` Better tests for panics
 * `[req]` Better tests for panics
 * `[spellcheck]` Better tests for panics
+* `[tmp]` Better tests for panics
 
 ### 12.60.1